+++ src/withComponents/WidgetProvider/widgetReducer.ts
@@ -6,6 +6,7 @@ export interface IWidgetProperties {
   id: string;
   hyperlink: string;
   participantSid: string;
+  title: string;
 }
 // `IWidgetState` is the state maintained by this context provider representing the widgets present in the room.
 export interface IWidgetState {